Moved all AST node semantic data and type data interfaces to new submodules in `@kipper/core`
