+++ kipper/core/src/compiler/semantics/language/statements.ts
@@ -25,6 +25,7 @@ import type { Expression } from "./expressions";
 import type { TargetASTNodeCodeGenerator, TargetASTNodeSemanticAnalyser } from "../../target-presets";
 import { LocalScope } from "../../local-scope";
 import { KipperNotImplementedError, UnableToDetermineSemanticDataError } from "../../../errors";
+import { IfStatementSemantics } from "../semantic-data";
 
 /**
  * Every antlr4 statement ctx type
@@ -185,31 +186,6 @@ export class CompoundStatement extends Statement<NoSemantics, NoTypeSemantics> {
 	targetSemanticAnalysis: TargetASTNodeSemanticAnalyser<CompoundStatement> = this.semanticAnalyser.compoundStatement;
 	targetCodeGenerator: TargetASTNodeCodeGenerator<CompoundStatement, Array<TranslatedCodeLine>> =
 		this.codeGenerator.compoundStatement;
-}
-
-/**
- * Semantics for AST Node {@link IfStatement}.
- * @since 0.9.0
- */
-export interface IfStatementSemantics extends SemanticData {
-	/**
-	 * The condition of the if-statement.
-	 * @since 0.9.0
-	 */
-	condition: Expression<any, any>;
-	/**
-	 * The body of the if-statement.
-	 * @since 0.9.0
-	 */
-	statementBody: Statement<any, any>;
-	/**
-	 * The alternative branch of the if-statement, which is optional. This alternative branch can either be:
-	 * - An else branch, if the type is a regular {@link Statement} (the statement that should be
-	 * evaluated in the else branch)
-	 * - Or an else-if branch, if the type is another {@link IfStatement}.
-	 * @since 0.9.0
-	 */
-	alternativeBranch?: IfStatement | Statement<any, any>;
 }
 
 /**
