+++ kipper/core/src/compiler/semantics/semantic-data/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Semantic data definitions for all AST nodes.
+ * @author Luna Klatzer
+ * @copyright 2021-2022 Luna Klatzer
+ * @since 0.10.0
+ */
+export * from "./definitions";
+export * from "./expressions";
+export * from "./statements";
