+++ kipper/core/src/compiler/semantics/type-data/statements.ts
@@ -1,0 +1,10 @@
+/**
+ * Semantic type data definitions for all statement AST nodes.
+ * @author Luna Klatzer
+ * @copyright 2021-2022 Luna Klatzer
+ * @since 0.10.0
+ */
+
+// No type data required for currently implemented statements.
+// This will likely change in the future though.
+export {};
